RAG_v.1/requirements.txt

chromadb
sentence-transformers
groq
python-dotenv

